refactor: replace with styles.css
