+++ src/react/loom-app/new-column-button/index.tsx
@@ -1,4 +1,3 @@
-import { css } from "@emotion/react";
 import Button from "../../shared/button";
 import Icon from "../../shared/icon";
