+++ src/react/loom-app/option-bar/active-filter-bubble/index.tsx
@@ -1,0 +1,14 @@
+import "./styles.css";
+
+interface Props {
+	numActive: number;
+}
+
+export default function ActiveFilterBubble({ numActive }: Props) {
+	if (numActive === 0) return <></>;
+	return (
+		<div className="dataloom-active-filter-bubble">
+			{numActive} active filter{numActive > 1 ? "s" : ""}
+		</div>
+	);
+}
