+++ src/react/loom-app/option-bar/active-filter-bubble/styles.css
@@ -1,0 +1,7 @@
+.dataloom-active-filter-bubble {
+    border-radius: 8px;
+    padding: 2px 6px;
+    color: var(--text-on-accent);
+    border: 1px solid var(--background-modifier-border);
+    background-color: var(--color-accent);
+}
